requirements.txt
requests
pandas
schedule
numpy
python-binance 